scx_layered: Add cost accounting
Add cost accounting for layers to make weights work on the BPF side.
This is done at both the CPU level as well as globally. When a CPU
runs out of budget it acquires budget from the global context. If a
layer runs out of global budgets then all budgets are reset. Weight
handling is done by iterating over layers by their available budget.
Layers budgets are proportional to their weights.

+++ scheds/rust/scx_layered/src/bpf/timer.bpf.h
@@ -20,7 +20,7 @@ struct layered_timer {
 	// if set to 0 the timer will only be scheduled once
 	u64 interval_ns;
 	u64 init_flags;
-	u64 start_flags;
+	int start_flags;
 };
 
 enum layer_timer_callbacks {
